requirements.txt
python-dotenv
PyInquirer
pyinstaller
click
docker-py
dockerpty
pyyaml